remove volumesfield
